requirements.txt
Flask==1.0.3
gunicorn==19.9.0
numpy==1.15.4
Pillow==5.3.0
Jinja2==2.10
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp36-cp36m-linux_x86_64.whl

